refactor: extract shared Dir and move and group Pac-Man board files into components/pacman
