+++ rn/package.json
@@ -3,17 +3,17 @@
   "main": "expo-router/entry",
   "version": "0.1.0",
   "dependencies": {
-    "@expo/ui": "~57.0.9",
+    "@expo/ui": "~57.0.11",
     "expo": "~57.0.10",
-    "expo-auth-session": "~7.0.8",
+    "expo-auth-session": "~57.0.7",
     "expo-constants": "~57.0.8",
-    "expo-crypto": "~15.0.7",
+    "expo-crypto": "~57.0.1",
     "expo-device": "~57.0.1",
     "expo-font": "~57.0.1",
     "expo-glass-effect": "~57.0.1",
-    "expo-image": "~57.0.2",
-    "expo-linking": "~57.0.5",
-    "expo-router": "~57.0.11",
+    "expo-image": "~57.0.3",
+    "expo-linking": "~57.0.6",
+    "expo-router": "~57.0.13",
     "expo-status-bar": "~57.0.1",
     "expo-symbols": "~57.0.2",
     "expo-system-ui": "~57.0.2",
