+++ rn/src/game/dir.ts
@@ -1,0 +1,1 @@
+export type Dir = { x: number; y: number };
